🧹 Spring cleaning? What's that.
	modified:   src/events/guildMemberUpdate.ts
	modified:   src/index.ts
	modified:   src/modules/credits/creditRoles.ts
	modified:   src/modules/credits/events/presenceUpdate.ts
	modified:   src/modules/credits/index.ts
	modified:   src/modules/moderation/commands/info.ts
	modified:   src/modules/moderation/commands/unmute.ts
	modified:   src/modules/moderation/commands/warn.ts
	modified:   src/modules/moderation/commands/warns.ts
	modified:   src/modules/moderation/index.ts
	modified:   src/modules/moderation/short-infos.ts
	modified:   src/modules/status/index.ts
	modified:   src/modules/utility/commands/help.ts
	modified:   src/modules/utility/commands/search.ts
	modified:   src/util/moduleLoader.ts

+++ src/modules/moderation/index.ts
@@ -15,8 +15,7 @@ export let blacklistedWords: string[] = [];
 			.first()
 			.members.fetch(mute.userId)
 			.then(m => {
-				if (!m.roles.cache.has(roles.muted))
-					coll.findOneAndDelete({ userId: mute.userId });
+				if (!m.roles.cache.has(roles.muted)) coll.findOneAndDelete({ userId: mute.userId });
 			})
 			.catch(() => {});
